Final push to 2.0
- Final update of the EnumTrait and enum.php template to have enums that
result in much less boilerplate for the consuming applications to deal
with.

- Adds governance docs
