+++ SECURITY.md
@@ -1,0 +1,17 @@
+# Security Policy
+
+## Supported Versions
+
+This project currently supports the following versions. 
+
+| Version | Supported          | Anticipated EOL |
+| ------- | ------------------ |-----------------|
+| 2.0.x   | :white_check_mark: | _TBD_           |
+
+## Reporting a Vulnerability
+
+I take security vulnerabilities very seriously. If you find a security vulnerability in code that I have created 
+please **do not open a public issue** and instead contact me directly at [security@cspray.io](mailto:security@cspray.io).
+This email is hosted on a service using end-to-end encryption and I will respond within 24 hours of receiving any email 
+at this address. We will then discuss further on exactly what needs to be done to resolve the vulnerability to the satsifaction 
+of all parties.
