+++ resources/templates/enum.php
@@ -13,25 +13,15 @@ final class <?= $this->getEnumName() ?> implements Enum {
 
     use EnumTrait;
 
-    private $enumValue;
-
-    private function __construct(string $enumValue) {
-        $this->enumValue = $enumValue;
+<?php foreach($this->getEnumValues() as $enumValue): ?>
+    public static function <?= $enumValue ?>() : self {
+        return self::getSingleton(<?= var_export($enumValue, true) ?>);
     }
 
+<?php endforeach; ?>
     // It is imperative that if you add a new value post code generation you add the method name here!
     static protected function getAllowedValues() : array {
         return [<?php foreach($this->getEnumValues() as $enumValue): echo var_export($enumValue, true), ', '; endforeach; ?>];
     }
 
-    <?php foreach($this->getEnumValues() as $enumValue): ?>
-    public static function <?= $enumValue ?>() : self {
-        return self::getSingleton(<?= var_export($enumValue, true) ?>);
-    }
-
-    <?php endforeach; ?>
-
-    public function toString() : string {
-        return $this->enumValue;
-    }
 }
