+++ src/Enum.php
@@ -2,23 +2,52 @@
 
 namespace Cspray\Yape;
 
+use Cspray\Yape\Exception\InvalidArgumentException;
+
 /**
- * A marker interface that identifies a class as one that is intended to serve as an Enum.
+ * An interface that marks a class as being an Enum.
  *
- * If you use the provided code generator and specify a value that is an enum it MUST implement this interface. Any
- * yape-generated enum will implement this interface. If you build your own enums from scratch using the provided traits
- * it is not necessary to implement this interface.
+ * By default enums generated by this library will implement this interface. However, if your application does not
+ * actually typehint this interface and you are building up enums from scratch using the provided EnumTrait then it is
+ * not strictly necessary to implement this interface, though it is highly recommended.
  *
  * @package Cspray\Yape
  * @license See LICENSE in source root
  */
 interface Enum {
 
+    /**
+     * Return a list of enum values associated with this enum.
+     *
+     * @return self[]
+     */
     static public function values() : array;
 
-    static public function valueOf(string $enumName);
+    /**
+     * Return an enum value that matches the given $value or throw an exception if the value is not valid for the given
+     * enum.
+     *
+     * $value SHOULD be the name of the static method that generated the enum instance.
+     *
+     * @param string $value
+     * @return self
+     * @throws InvalidArgumentException
+     */
+    static public function valueOf(string $value);
 
+    /**
+     * Determines whether or not the given $compare value is strictly equal to the enum instance.
+     *
+     * @param $compare
+     * @return bool
+     */
     public function equals($compare) : bool;
 
+    /**
+     * Return the enum value for the given enum instance; the return value of this method should result in a valid enum
+     * being created from valueOf.
+     *
+     * @return string
+     */
     public function toString() : string;
 }
